Audit.NET.DynamoDB to net462;netstandard2.0;net6.0
