+++ src/Audit.NET.DynamoDB/Audit.NET.DynamoDB.csproj
@@ -5,7 +5,7 @@
     <Copyright>Copyright 2018</Copyright>
     <AssemblyTitle>Audit.DynamoDB</AssemblyTitle>
     <Authors>Federico Colombo</Authors>
-    <TargetFrameworks>netstandard2.0;net45;net5.0</TargetFrameworks>
+    <TargetFrameworks>net462;netstandard2.0;net6.0</TargetFrameworks>
     <DefineConstants>$(DefineConstants);STRONG_NAME</DefineConstants>
     <NoWarn>$(NoWarn);1591</NoWarn>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
@@ -31,7 +31,7 @@
   </ItemGroup>
 	
   <ItemGroup>
-	<PackageReference Include="AWSSDK.DynamoDBv2" Version="3.7.101.55" /> 
+	<PackageReference Include="AWSSDK.DynamoDBv2" Version="3.7.300.17" /> 
   </ItemGroup>
   
   <ItemGroup>
