+++ test/Audit.IntegrationTest/IntegrationTests.cs
@@ -1,9 +1,5 @@
-﻿using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DocumentModel;
-using Amazon.DynamoDBv2.Model;
-using Audit.Core;
+﻿using Audit.Core;
 using Audit.Core.Providers;
-using Audit.DynamoDB.Providers;
 using Audit.Kafka.Providers;
 using Audit.MongoDB.Providers;
 using Audit.SqlServer.Providers;
@@ -26,10 +22,6 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using System.Data.Common;
 using System.Threading;
-#endif
-#if NET5_0_OR_GREATER
-using Azure.Storage;
-using Azure.Storage.Blobs.Models;
 #endif
 #if NK_JSON
 using Newtonsoft.Json;
@@ -489,56 +481,6 @@ namespace Audit.IntegrationTest
             }
 
             [Test]
-            [Category("Dynamo")]
-            public void TestStressDynamo()
-            {
-                int N = 32;
-                SetDynamoSettings();
-                var hashes = new HashSet<string>();
-                int count = 0;
-                Audit.Core.Configuration.AddCustomAction(ActionType.OnEventSaved, scope =>
-                {
-                    count++;
-                    hashes.Add((scope.EventId as object[])[0].ToString());
-                });
-
-                var rnd = new Random();
-
-                //Parallel random insert into event1, event2 and event3 containers
-                Parallel.ForEach(Enumerable.Range(1, N), i =>
-                {
-                    var eventType = "AuditEvents";
-                    var x = "start";
-                    using (var s = new AuditScopeFactory().Create(eventType, () => x, EventCreationPolicy.InsertOnStartReplaceOnEnd, null))
-                    {
-                        x = "end";
-                    }
-                });
-
-                Assert.AreEqual(N, hashes.Count);
-                Assert.AreEqual(N*2, count);
-
-                // Assert events
-                int maxCheck = N / 4;
-                int check = 0;
-                foreach(var hash in hashes)
-                {
-                    if (check++ > maxCheck)
-                    {
-                        break;
-                    }
-                    var ddp = (Configuration.DataProvider as DynamoDataProvider);
-                    var ev = ddp.GetEvent<AuditEvent>((Primitive)hash, (Primitive)DateTime.Now.Year);
-
-                    Assert.NotNull(ev);
-                    Assert.AreEqual("AuditEvents", ev.EventType);
-                    Assert.AreEqual(DateTime.Now.Year.ToString(), ev.CustomFields["SortKey"].ToString());
-                    Assert.AreEqual(hash, ev.CustomFields["HashKey"].ToString());
-                }
-            }
-
-
-            [Test]
             [Category("AzureBlob")]
             public void TestAzureTable()
             {
@@ -697,24 +639,6 @@ namespace Audit.IntegrationTest
                 await TestUpdateAsync();
             }
 #endif
-
-            [Test]
-            [Category("Dynamo")]
-            public void TestDynamo()
-            {
-                SetDynamoSettings();
-                TestUpdate();
-                TestInsert();
-                TestDelete();
-            }
-
-            [Test]
-            [Category("Dynamo")]
-            public async Task TestDynamoAsync()
-            {
-                SetDynamoSettings();
-                await TestUpdateAsync();
-            }
 
             public struct TestStruct
             {
@@ -1139,53 +1063,6 @@ namespace Audit.IntegrationTest
 
             }
 #endif
-            public void SetDynamoSettings()
-            {
-                var url = "http://localhost:8000";
-                var tableName = "AuditEvents";
-                CreateDynamoTable(tableName).GetAwaiter().GetResult();
-
-                Audit.Core.Configuration.Setup()
-                    .UseDynamoDB(config => config
-                        .UseUrl(url)
-                        .Table(tableName)
-                        .SetAttribute("HashKey", ev => Guid.NewGuid())
-                        .SetAttribute("SortKey", ev => ev.StartDate.Year))
-                    .WithCreationPolicy(EventCreationPolicy.InsertOnStartReplaceOnEnd)
-                    .ResetActions();
-            }
-
-            private async Task CreateDynamoTable(string tableName)
-            {
-                AmazonDynamoDBConfig ddbConfig = new AmazonDynamoDBConfig
-                {
-                    ServiceURL = "http://localhost:8000"
-                };
-                var client = new AmazonDynamoDBClient(ddbConfig);
-                try
-                {
-                    await client.DeleteTableAsync(tableName);
-                }
-                catch
-                {
-                }
-
-                await client.CreateTableAsync(new CreateTableRequest()
-                {
-                    TableName = tableName,
-                    KeySchema = new List<KeySchemaElement>()
-                    {
-                        new KeySchemaElement("HashKey", KeyType.HASH),
-                        new KeySchemaElement("SortKey", KeyType.RANGE)
-                    },
-                    AttributeDefinitions = new List<AttributeDefinition>()
-                    {
-                        new AttributeDefinition("HashKey", ScalarAttributeType.S),
-                        new AttributeDefinition("SortKey", ScalarAttributeType.N)
-                    },
-                    ProvisionedThroughput = new ProvisionedThroughput(100, 100)
-                });
-            }
 
             public static void ExecuteStoredProcedure(IntegrationTests.CustomerOrder order, IntegrationTests.OrderStatus status)
             {
